requirements.txt
aiohttp==3.8.4
openai==1.51.2
pydantic==1.9.2
python-dotenv==1.0.1
together==1.3.2
